only show "Delete" option when code can actually be deleted
fixes #12

+++ editor/src/code_editor_renderer.rs
@@ -796,15 +796,19 @@ impl<'a, T: UiToolkit> CodeEditorRenderer<'a, T> {
                                    })
                 },
                 &|| {
-                    let cmd_buffer = Rc::clone(&self.command_buffer);
-                    self.ui_toolkit.draw_menu_item("Delete", move || {
-                                       cmd_buffer.borrow_mut()
-                                           .add_editor_command(move |editor| {
-                                               editor.delete_node_id(code_node_id_to_act_on);
-                                           })
-                                   })
+                    if self.code_editor.can_be_deleted(code_node_id_to_act_on) {
+                        let cmd_buffer = Rc::clone(&self.command_buffer);
+                        self.ui_toolkit.draw_menu_item("Delete", move || {
+                            cmd_buffer.borrow_mut()
+                                .add_editor_command(move |editor| {
+                                    editor.delete_node_id(code_node_id_to_act_on);
+                                })
+                        })
+                    } else {
+                        self.ui_toolkit.draw_all(&[])
+                    }
                 },
-            ])
+                           ])
                        })
     }
